+++ src/extension.ts
@@ -6,83 +6,108 @@ import * as vscode from 'vscode';
 // Your extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext) {
 	context.subscriptions.push(vscode.commands.registerCommand("qqol.backspaceLeft", backspace));
+
+	const config = vscode.workspace.getConfiguration("qqol");
+	if (config.get("trimIndentOnlyLinesOnSave")) {
+		context.subscriptions.push(vscode.workspace.onWillSaveTextDocument(trimSpaceOnEmptyLines));
+	}
 }
 
-function areMatchingParens(start: string, end: string): boolean {
-	return (start === "(" && end === ")") || (start === "{" && end === "}") ||
-		(start === "[" && end === "]") || (start === "<" && end === ">") ||
-		(start === "\"" && end === "\"") || (start === "'" && end === "'");
+function trimSpaceOnEmptyLines(event: vscode.TextDocumentWillSaveEvent) {
+	const editor = vscode.window.activeTextEditor;
+	if (!editor || editor.document !== event.document) {
+		return;
+	}
+
+	const document = editor.document;
+	const edits = [];
+
+	const reg = /^\s+$/;
+
+	for (let i = 0; i < document.lineCount; ++i) {
+		const line = document.lineAt(i);
+		if (reg.test(line.text)) {
+			edits.push(vscode.TextEdit.delete(line.range));
+		}
+	}
+
+	if (edits.length !== 0) {
+		event.waitUntil(Promise.resolve(edits));
+	}
 }
 
-function backspace() {
-	let editor = vscode.window.activeTextEditor;
+async function backspace() {
+	const editor = vscode.window.activeTextEditor;
 	if (!editor) {
 		return;
 	}
 
-	const tabSize = editor.options.tabSize as number;
-	const selections = editor.selections;
+	const selections = [...editor.selections].sort((a, b) => b.active.line - a.active.line);
 	const document = editor.document;
 
-	const newSelections = selections.map(selection => {
+	const removedRanges = [];
+
+	for (const selection of selections) {
 		if (!selection.isEmpty) {
-			return selection;
+			continue;
 		}
 
-		const start = selection.start;
-		const line = document.lineAt(start);
-		if (start.line === 0) {
-			const newStart = start.character === 0 ?
-				new vscode.Position(0, 0) :
-				new vscode.Position(start.line, start.character - 1);
-			selection = new vscode.Selection(newStart, selection.end);
-			
-			if (selection.end.character > 0 && selection.end.character < line.text.length &&
-				areMatchingParens(line.text[selection.end.character - 1], line.text[selection.end.character])) {
-				const newEnd = selection.end.translate(0, 1);
-				selection = new vscode.Selection(selection.start, newEnd);
-			}
-			return selection;
+		const pos = selection.active;
+		if (pos.line === 0) {
+			continue;
 		}
 
-		const prevLine = document.lineAt(start.line - 1);
-		
-		const newStart = start.character === 0 ?
-			new vscode.Position(start.line - 1, prevLine.text.length) :
-			new vscode.Position(start.line, start.character - 1);
-		selection = new vscode.Selection(newStart, selection.end);
-		
-		if (selection.end.character > 0 && selection.end.character < line.text.length &&
-			areMatchingParens(line.text[selection.end.character - 1], line.text[selection.end.character])) {
-			const newEnd = selection.end.translate(0, 1);
-			selection = new vscode.Selection(selection.start, newEnd);
-		}
-		
-		const prevFirstNonWhitespace = prevLine.firstNonWhitespaceCharacterIndex;
-		if (line.firstNonWhitespaceCharacterIndex < start.character) {
-			return selection;
-		} else if (prevFirstNonWhitespace !== 0 && start.character > prevFirstNonWhitespace) {
-			const anchor = new vscode.Position(start.line, prevFirstNonWhitespace);
-			return new vscode.Selection(anchor, start);
-		}
-		
-		const text = line.text.substring(0, start.character);
-		const match = /^[\t ]+$/.exec(text);
-		if (match) {
-			const anchor = prevLine.range.end;
-			return new vscode.Selection(anchor, start);
-		}
+		const line = document.lineAt(pos.line);
+		const indentEnd = line.firstNonWhitespaceCharacterIndex;
 
-		return selection;
-	});
-	
-	const returned = editor.edit(editorBuilder => {
-		newSelections.forEach(deletion => {
-			editorBuilder.delete(deletion);
-		});
-	});
+		if (pos.character <= indentEnd) {
+			const prevLine = document.lineAt(pos.line - 1);
+			removedRanges.push(new vscode.Range(prevLine.range.end, pos));
+		}
+	}
 
-	return returned;
+	if (removedRanges.length === 0) {
+		await vscode.commands.executeCommand("deleteLeft");
+		return;
+	}
+
+	removedRanges.sort((a, b) => b.start.line - a.start.line);
+	const filteredRanges: vscode.Range[] = [];
+	let lastRangeStart = Infinity;
+	for (const removed of removedRanges) {
+		if (removed.end.line < lastRangeStart) {
+			filteredRanges.push(removed);
+			lastRangeStart = removed.start.line;
+		}
+	}
+
+	await editor.edit(editBuilder => {
+		for (const filtered of filteredRanges) {
+			editBuilder.delete(filtered);
+		}
+	},
+	{ undoStopBefore: true, undoStopAfter: false });
+
+	await vscode.commands.executeCommand("editor.action.reindentlines");
+
+	const newSelections: vscode.Selection[] = [];
+	const set = new Set();
+
+	for (const selection of selections) {
+		const oldLine = selection.active.line;
+		const numRemovedBefore = filteredRanges.filter(range => range.start.line <= oldLine).length;
+		const newLine = Math.max(oldLine - numRemovedBefore, 0);
+		const actual = Math.min(newLine, document.lineCount - 1);
+		const line = document.lineAt(actual);
+		const newPos = new vscode.Position(actual, line.range.end.character);
+		const key = `${newPos.line}:${newPos.character}`;
+		if (!set.has(key)) {
+			set.add(key);
+			newSelections.push(new vscode.Selection(newPos, newPos));
+		}
+	}
+
+	editor.selections = newSelections;
 }
 
 // This method is called when your extension is deactivated